Domain management: Hook purchases in bulk tables
