+++ client/my-sites/domains/domain-management/list/bulk-all-domains.tsx
@@ -1,22 +1,15 @@
 import { isEnabled } from '@automattic/calypso-config';
-import {
-	DomainsTable,
-	useDomainsTable,
-	DomainStatusPurchaseActions,
-	ResponseDomain,
-} from '@automattic/domains-table';
+import { DomainsTable, useDomainsTable } from '@automattic/domains-table';
 import { useTranslate } from 'i18n-calypso';
 import { UsePresalesChat } from 'calypso/components/data/domain-management';
 import InlineSupportLink from 'calypso/components/inline-support-link';
 import Main from 'calypso/components/main';
 import BodySectionCssClass from 'calypso/layout/body-section-css-class';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
-import { handleRenewNowClick, shouldRenderExpiringCreditCard } from 'calypso/lib/purchases';
-import { Purchase } from 'calypso/lib/purchases/types';
 import { useOdieAssistantContext } from 'calypso/odie/context';
-import { useDispatch } from 'calypso/state';
 import DomainHeader from '../components/domain-header';
 import OptionsDomainButton from './options-domain-button';
+import { usePurchaseActions } from './use-purchase-actions';
 
 interface BulkAllDomainsProps {
 	analyticsPath: string;
@@ -25,7 +18,6 @@ interface BulkAllDomainsProps {
 
 export default function BulkAllDomains( props: BulkAllDomainsProps ) {
 	const { domains, isLoading } = useDomainsTable();
-	const dispatch = useDispatch();
 	const translate = useTranslate();
 	const { sendNudge } = useOdieAssistantContext();
 
@@ -53,36 +45,7 @@ export default function BulkAllDomains( props: BulkAllDomainsProps ) {
 		<OptionsDomainButton key="breadcrumb_button_1" specificSiteActions allDomainsList />,
 	];
 
-	const isCreditCardExpiring = ( purchases: Purchase[] ) => ( domain: ResponseDomain ) => {
-		const purchase = purchases?.find(
-			( p ) => p.id === parseInt( domain.subscriptionId ?? '', 10 )
-		);
-
-		return ( purchase && dispatch( shouldRenderExpiringCreditCard( purchase ) ) ) ?? false;
-	};
-
-	const isPurchasedDomain = ( purchases: Purchase[] ) => ( domain: ResponseDomain ) => {
-		const purchase = purchases?.find(
-			( p ) => p.id === parseInt( domain.subscriptionId ?? '', 10 )
-		);
-		return !! purchase;
-	};
-
-	const onRenewNowClick =
-		( purchases: Purchase[] ) => ( siteSlug: string, domain: ResponseDomain ) => {
-			const purchase = purchases?.find(
-				( p ) => p.id === parseInt( domain.subscriptionId ?? '', 10 )
-			);
-			if ( purchase ) {
-				dispatch( handleRenewNowClick( purchase, siteSlug ) );
-			}
-		};
-
-	const purchaseActions: DomainStatusPurchaseActions = {
-		isCreditCardExpiring: isCreditCardExpiring( [] ),
-		isPurchasedDomain: isPurchasedDomain( [] ),
-		onRenewNowClick: onRenewNowClick( [] ),
-	};
+	const purchaseActions = usePurchaseActions();
 
 	return (
 		<>
